+++ cpmpy/tools/explain/__init__.py
@@ -1,3 +1,4 @@
 from .mus import *
 from .mss import *
 from .mcs import *
+from .marco import *
